+++ src/plugin.js
@@ -1,7 +1,16 @@
+import postcss from 'postcss'
+import lightningcss, { Features } from 'lightningcss'
+import browserslist from 'browserslist'
 import setupTrackingContext from './lib/setupTrackingContext'
 import processTailwindFeatures from './processTailwindFeatures'
 import { env } from './lib/sharedState'
 import { findAtConfigPath } from './lib/findAtConfigPath'
+import { handleImportAtRules } from './lib/handleImportAtRules'
+import { version as tailwindVersion } from '../package.json'
+
+function license() {
+  return `/* ! tailwindcss v${tailwindVersion} | MIT License | https://tailwindcss.com */\n`
+}
 
 module.exports = function tailwindcss(configOrPath) {
   return {
@@ -13,6 +22,7 @@ module.exports = function tailwindcss(configOrPath) {
           console.time('JIT TOTAL')
           return root
         },
+      ...handleImportAtRules(),
       function (root, result) {
         // Use the path for the `@config` directive if it exists, otherwise use the
         // path for the file being processed
@@ -34,66 +44,82 @@ module.exports = function tailwindcss(configOrPath) {
 
         processTailwindFeatures(context)(root, result)
       },
-      __OXIDE__ &&
-        function lightningCssPlugin(_root, result) {
-          let postcss = require('postcss')
-          let lightningcss = require('lightningcss')
-          let browserslist = require('browserslist')
+      function lightningCssPlugin(_root, result) {
+        let map = result.map ?? result.opts.map
 
-          try {
-            let transformed = lightningcss.transform({
-              filename: result.opts.from,
-              code: Buffer.from(result.root.toString()),
-              minify: false,
-              sourceMap: !!result.map,
-              inputSourceMap: result.map ? result.map.toString() : undefined,
-              targets:
-                typeof process !== 'undefined' && process.env.JEST_WORKER_ID
-                  ? { chrome: 106 << 16 }
-                  : lightningcss.browserslistToTargets(
-                      browserslist(require('../package.json').browserslist)
-                    ),
+        let intermediateResult = result.root.toResult({
+          map: map ? { inline: true } : false,
+        })
 
-              drafts: {
-                nesting: true,
-                customMedia: true,
-              },
-            })
+        let intermediateMap = intermediateResult.map?.toJSON?.() ?? map
 
-            result.map = Object.assign(result.map ?? {}, {
-              toJSON() {
-                return transformed.map.toJSON()
-              },
-              toString() {
-                return transformed.map.toString()
-              },
-            })
+        try {
+          let resolvedBrowsersListConfig = browserslist.findConfig(
+            result.opts.from ?? process.cwd()
+          )?.defaults
+          let defaultBrowsersListConfig = require('../package.json').browserslist
+          let browsersListConfig = resolvedBrowsersListConfig ?? defaultBrowsersListConfig
 
-            result.root = postcss.parse(transformed.code.toString('utf8'))
-          } catch (err) {
-            if (typeof process !== 'undefined' && process.env.JEST_WORKER_ID) {
-              let lines = err.source.split('\n')
-              err = new Error(
-                [
-                  'Error formatting using Lightning CSS:',
-                  '',
-                  ...[
-                    '```css',
-                    ...lines.slice(Math.max(err.loc.line - 3, 0), err.loc.line),
-                    ' '.repeat(err.loc.column - 1) + '^-- ' + err.toString(),
-                    ...lines.slice(err.loc.line, err.loc.line + 2),
-                    '```',
-                  ],
-                ].join('\n')
-              )
+          let transformed = lightningcss.transform({
+            filename: result.opts.from,
+            code: Buffer.from(intermediateResult.css),
+            minify: false,
+            sourceMap: !!intermediateMap,
+            targets: lightningcss.browserslistToTargets(browserslist(browsersListConfig)),
+            drafts: {
+              nesting: true,
+              customMedia: true,
+            },
+            nonStandard: {
+              deepSelectorCombinator: true,
+            },
+            include: Features.Nesting,
+            exclude: Features.LogicalProperties,
+          })
+
+          let code = transformed.code.toString()
+
+          // https://postcss.org/api/#sourcemapoptions
+          if (intermediateMap && transformed.map != null) {
+            let prev = transformed.map.toString()
+
+            if (typeof intermediateMap === 'object') {
+              intermediateMap.prev = prev
+            } else {
+              code = `${code}\n/*# sourceMappingURL=data:application/json;base64,${Buffer.from(
+                prev
+              ).toString('base64')} */`
             }
-
-            if (Error.captureStackTrace) {
-              Error.captureStackTrace(err, lightningCssPlugin)
-            }
-            throw err
           }
-        },
+
+          result.root = postcss.parse(license() + code, {
+            ...result.opts,
+            map: intermediateMap,
+          })
+        } catch (err) {
+          if (err.source && typeof process !== 'undefined' && process.env.JEST_WORKER_ID) {
+            let lines = err.source.split('\n')
+            err = new Error(
+              [
+                'Error formatting using Lightning CSS:',
+                '',
+                ...[
+                  '```css',
+                  ...lines.slice(Math.max(err.loc.line - 3, 0), err.loc.line),
+                  ' '.repeat(err.loc.column - 1) + '^-- ' + err.toString(),
+                  ...lines.slice(err.loc.line, err.loc.line + 2),
+                  '```',
+                ],
+              ].join('\n')
+            )
+          }
+
+          if (Error.captureStackTrace) {
+            Error.captureStackTrace(err, lightningCssPlugin)
+          }
+          throw err
+        }
+      },
       env.DEBUG &&
         function (root) {
           console.timeEnd('JIT TOTAL')